fix(input): use uint raw device change constants

+++ src/LatencyPilot.Platform.Windows/Interop/User32RawInputTiming.cs
@@ -51,8 +51,8 @@ internal static partial class User32RawInput
     internal const uint RemoveFlag = 0x00000001;
     internal const uint WmInputDeviceChange = 0x00FE;
     internal const uint WmInput = 0x00FF;
-    internal const nuint GidcArrival = 1;
-    internal const nuint GidcRemoval = 2;
+    internal const uint GidcArrival = 1;
+    internal const uint GidcRemoval = 2;
     internal const uint WmApp = 0x8000;
     internal const int WindowProcedureIndex = -4;
     internal static readonly nint MessageOnlyWindow = new(-3);
